STK-01: Separando as lógicas em um serviço.
